simplify execution of and provide examples
It's better to provide environment variables for configuring containers.

The new entrypoint adds support for setting binary arguments based on
environment variables.

Additionally, a default server settings configuration and maplist for bf2
and special forces which were added to the examples directory.

And for easier deployment an example docker compose file was added
showing how to run a vanilla bf2 as well as the special forces mod.